api-gateway/requirements.txt
fastapi
uvicorn
requests
python-multipart
pydantic
python-jose[cryptography]
passlib[bcrypt]